moved the pipeline to its own header
