+++ space_assault/src/app_main.cpp
@@ -1,6 +1,5 @@
 #include "game/game.hpp"
 namespace sa = assault::game;
-#include "lodepng/lodepng.h"
 
 int main()
 {
